improved camera movement

+++ src/camera.h
@@ -39,7 +39,7 @@ public:
     void RecalculateView();
 public:
     int type = 0;
-    int horizontalFOV = 90;
+    float horizontalFOV = 90.0f;
     float damping = 0.9f;
     float focal_length = 4.0f;
     float aperture = 0.0f;
